Predictive Analysis from Car Accident 2021
It's a project that I'll be evolving during the days. This one is my first public project, so I believe that could shows what I want it, for example, I cover EDA, Data pre-processing and implement of ML algorithms. This' the first version to show how I worked with a Kaggle dataset that's truthly interesting and we can aproache more than what we had.